LRAC-13369 osb-faro-contacts-service:Rename
